+++ src/lib/Components/Inbox/Conversations/__stories__/ArtworkPreview.story.tsx
@@ -23,5 +23,3 @@
 //     artwork.title = "Canisters Canisters Canisters Canisters Canisters Canisters"
 //     return <StubContainer Component={ArtworkPreview} props={{ artwork }} />
 //   })
-
-export default undefined